Do not use custom deleter for canceler action
shared_ptr deleter is owned by control block and is not destroyed until
write end of the future (promise/packaged_task/...) is destroyed. This
behaviour is confusing and differs from the rest of the library: every
user provided function object which must only be called once is guarantied
to be deleted right after call.

Closes #15