Change behavior of [.stop()] and [.wait()] on [pc::static_thread_pool]
[pc::static_thread_pool] tends to follow https://wg21.link/p0443 specification and it
looks like current implementation misses the difference between [.stop()] and [.wait()] methods.

> `void stop();`
>
>  Effects: Signals the threads in the pool to complete as soon as possible.
>  If a thread is currently executing a function object, the thread will exit
>  only after completion of that function object. Invocation of stop() returns
>  without waiting for the threads to complete. Subsequent invocations to attach
>  complete immediately.

>  `void wait();`
>
>   Effects: If not already stopped, signals the threads in the pool to complete
>   once the outstanding work is 0.

"outstanding work" is defined in the document to include task which are scheduled
but not yet started. Current implementation stops receiving incoming tasks and
discards not yet started tasks and this is a bug, which this commit strives to fix.